Hoist the field layout boilerplate from MetadataType to DefType since all of the code specific to a particular kind of layout is encapsulated by the FiledLayoutAlgorithm
